pyproject.toml
[project]
name = "realestate-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.99.6",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "redis>=6.4.0",
    "streamlit>=1.48.0",
]
